OrlandoDiaz_Project1
The beginning of my Project is just the sketch of the character I’ll be
using throughout the project. Look up “Bill Cipher” on Google for a
picture of the character in action in the cartoon. I tried to replicate
him the best I could. (Still working on the limbs though which are the
most difficult part)