GROW
This is not exactly an adjective, but I think that it is a word that I
can work with. It is not much now because I am still figuring out how to
execute what I have in mind. I want the letters to expand as the mouse
moves around but so far is hasn't worked out.