Updated Project 1
I was going to push an updated version of my project on Wednesday but
decided to change my adjective, which is why I haven’t push anything
until now. I decided to change it to paranoid. However, I am having
trouble trying to get the body to repel the mouse.